Better waf venv support

+++ waf/virtualenv.py
@@ -47,11 +47,12 @@ class venv_link_task(Task.Task):
       sdirnode = self.generator.path.get_bld()
       sdirnode.mkdir()
       for snode in self.inputs:
-          dnode = snode.get_bld()
-          if not os.path.isdir(dnode.parent.abspath()):
-              dnode.parent.mkdir()
-          if not os.path.exists(dnode.abspath()):
-              os.symlink(os.path.relpath(snode.abspath(), os.path.join(dnode.abspath(), "..")), dnode.abspath())
+          if not os.path.exists(snode.abspath()):
+              dnode = snode.get_bld()
+              if not os.path.isdir(dnode.parent.abspath()):
+                  dnode.parent.mkdir()
+              if not os.path.exists(dnode.abspath()):
+                  os.symlink(os.path.relpath(snode.abspath(), os.path.join(dnode.abspath(), "..")), dnode.abspath())
       initnode = sdirnode.find_or_declare('__init__.py')
       if not os.path.exists(initnode.abspath()):
           initnode.write("")
@@ -64,16 +65,21 @@ class venv_link_task(Task.Task):
 @TaskGen.before('process_source', 'process_rule')
 @TaskGen.feature('venv_package')
 def venv_package(self):
-    srclist = []
+    initnode = self.path.find_or_declare('__init__.py')
+    if not os.path.exists(initnode.abspath()):
+        initnode.write("")
+    srclist = [initnode]
     for src in Utils.to_list(getattr(self, "pysource", [])):
         if isinstance(src, str):
             snode = self.path.find_node(src)
         else:
             snode = src
-        srclist.append(snode)
+        if not initnode in srclist:
+            srclist.append(snode)
     self.env["VENV_PATH"] = os.path.join(self.bld.bldnode.abspath(), ".conf_check_venv")
-    dst = self.bld.bldnode.find_node(".conf_check_venv")
-    links = self.create_task('venv_link', src=srclist, tgt=dst)
+    snode = self.path.abspath()
+    dnode = self.bld.bldnode.find_dir(os.path.join("lib", ("python%s" % self.env.PYTHON_VERSION), "site-packages", os.path.basename(snode)))
+    links = self.create_task('venv_link', src=list(srclist), tgt=dnode)
 
 def configure(self):
     try:
